requirements.txt
python-dotenv
chainlit
openai
aiosqlite
bcrypt
aiohttp
langchain-openai
